Reorganize ducks/tableMetadata + some cleanup
