+++ amundsen_application/static/js/components/SearchPage/index.tsx
@@ -8,15 +8,15 @@ import SearchList from './SearchList';
 import InfoButton from '../common/InfoButton';
 import { TableResource } from "../common/ResourceListItem/types";
 
-import { ExecuteSearchRequest } from '../../ducks/search/types';
-import { GetPopularTablesRequest } from '../../ducks/popularTables/types';
-// TODO: Use css-modules instead of 'import'
-import './styles.scss';
 import {
+  ExecuteSearchRequest,
   DashboardSearchResults,
   TableSearchResults,
   UserSearchResults
 } from "../../ducks/search/types";
+import { GetPopularTablesRequest } from '../../ducks/popularTables/types';
+// TODO: Use css-modules instead of 'import'
+import './styles.scss';
 
 const RESULTS_PER_PAGE = 10;
