+++ amundsen_application/static/js/components/Tags/TagInput/index.tsx
@@ -7,7 +7,7 @@ import CreatableSelect from 'react-select/lib/Creatable';
 import makeAnimated from 'react-select/lib/animated';
 
 import { GetAllTagsRequest } from '../../../ducks/allTags/types';
-import { UpdateTagsRequest } from '../../../ducks/tableMetadata/tags/reducer';
+import { UpdateTagsRequest } from '../../../ducks/tableMetadata/types';
 
 import TagInfo from "../TagInfo";
 import { Tag, UpdateTagMethod, UpdateTagData } from '../types';
