+++ amundsen_application/static/js/ducks/allTags/api/v0.ts
@@ -1,11 +1,11 @@
 import axios, { AxiosResponse, AxiosError } from 'axios';
 
-import { AllTagsPayload } from '../types';
+import { AllTagsResponse } from '../types';
 
 import { sortTagsAlphabetical } from '../../utilMethods';
 
 export function metadataAllTags() {
-  return axios.get('/api/metadata/v0/tags').then((response: AxiosResponse<AllTagsPayload>) => {
+  return axios.get('/api/metadata/v0/tags').then((response: AxiosResponse<AllTagsResponse>) => {
     return response.data.tags.sort(sortTagsAlphabetical);
   })
   .catch((error: AxiosError) => {
