+++ amundsen_application/static/js/ducks/allTags/sagas.ts
@@ -1,13 +1,9 @@
 import { call, put, takeEvery } from 'redux-saga/effects';
 import { SagaIterator } from 'redux-saga';
 
-import {
-  GetAllTags,
-} from './types';
+import { GetAllTags } from './types';
 
-import {
-  metadataAllTags,
-} from './api/v0';
+import { metadataAllTags } from './api/v0';
 
 export function* getAllTagsWorker(): SagaIterator {
   try {
