+++ amundsen_application/static/js/ducks/announcements/api/v0.ts
@@ -1,13 +1,13 @@
 import axios, { AxiosResponse, AxiosError } from 'axios';
 
-import { AnnouncementsPayload } from '../types';
+import { AnnouncementsResponse } from '../types';
 
 export function announcementsGet() {
   return axios({
       method: 'get',
       url: '/api/announcements/v0/',
     })
-    .then((response: AxiosResponse<AnnouncementsPayload>) => {
+    .then((response: AxiosResponse<AnnouncementsResponse>) => {
       return response.data.posts;
     })
     .catch((error: AxiosError) => {
