+++ amundsen_application/static/js/ducks/search/api/v0.ts
@@ -1,20 +1,8 @@
 import axios, { AxiosResponse, AxiosError } from 'axios';
-import {
-  DashboardSearchResults,
-  TableSearchResults,
-  UserSearchResults,
-} from '../types';
+
+import { SearchResponse } from '../types';
 
 import { SearchReducerState } from '../reducer';
-
-interface SearchResponse {
-  msg: string;
-  status_code: number;
-  search_term: string;
-  dashboards: DashboardSearchResults;
-  tables: TableSearchResults;
-  users: UserSearchResults;
-}
 
 function transformSearchResults(data: SearchResponse): SearchReducerState {
   return {
