+++ amundsen_application/static/js/ducks/tableMetadata/api/helpers.ts
@@ -1,11 +1,11 @@
-/** TODO: We will introduce better typing for function parameters return types */
+import { GetTableDataRequest, TableMetadata, TableDataResponse, Tag, User } from '../../tableMetadata/types';
 
 import { filterFromObj, sortTagsAlphabetical } from '../../utilMethods';
 
 /**
  * Generates the query string parameters needed for requests that act on a particular table resource.
  */
-export function getTableParams(tableDataObject) {
+export function getTableParams(tableDataObject: TableMetadata | GetTableDataRequest): string {
   const { cluster, database, schema, table_name } = tableDataObject;
   return `db=${database}&cluster=${cluster}&schema=${schema}&table=${table_name}`;
 }
@@ -13,17 +13,17 @@ export function getTableParams(tableDataObject) {
 /**
  * Parses the response for table metadata to create a TableMetadata object
  */
-export function getTableDataFromResponseData(responseData) {
-  return filterFromObj(responseData, ['owners', 'tags']);
+export function getTableDataFromResponseData(responseData: TableDataResponse): TableMetadata {
+  return filterFromObj(responseData.tableData, ['owners', 'tags']) as TableMetadata;
 }
 
 /**
  * Parses the response for table metadata to return the array of table owners
  */
-export function getTableOwnersFromResponseData(responseData) {
+export function getTableOwnersFromResponseData(responseData: TableDataResponse): { [id: string] : User } {
   // TODO: owner needs proper id, until then we have to remember that we are using display_name
-  const ownerObj = responseData.owners.reduce((resultObj, currentOwner) => {
-    resultObj[currentOwner.display_name] = currentOwner;
+  const ownerObj = responseData.tableData.owners.reduce((resultObj, currentOwner) => {
+    resultObj[currentOwner.display_name] = currentOwner as User;
     return resultObj;
   }, {});
   return ownerObj;
@@ -32,6 +32,6 @@ export function getTableOwnersFromResponseData(responseData) {
 /**
  * Parses the response for table metadata to return an array of sorted table tags
  */
-export function getTableTagsFromResponseData(responseData) {
-  return responseData.tags.sort(sortTagsAlphabetical);
+export function getTableTagsFromResponseData(responseData: TableDataResponse): Tag[] {
+  return responseData.tableData.tags.sort(sortTagsAlphabetical);
 }
